soc: arch: stm32g0: enable interrupt sharing for ucpd1/ucpd2/usb
Enable the shared interrupt muxer for ucpd1/2 devices on stm32g071/81
and ucpd1/ucpd2/usb devices on stm32g0b1/c1 in the device tree, and
configure the multilevel interrupt framework accordingly.

For that to the interrupt sharing support in the UCPD driver needs to be
removed as the instances won't see the interrupt of each other anymore.

Signed-off-by: Aurelien Jarno <aurelien@aurel32.net>